MAINT: moved all IOptimizer-related classes to separate file; removed redundancy dependencies; tests updates
