+++ itwm_example/mco/mco.py
@@ -2,79 +2,16 @@ import logging
 import sys
 
 import numpy as np
-from scipy import optimize
 
-from traits.api import (
-    HasStrictTraits,
-    List,
-    Float,
-    Instance,
-    Interface,
-    provides,
-    Type,
-)
+from traits.api import Type
 
-from force_bdss.api import BaseMCO, BaseMCOParameter, DataValue
-from force_bdss.mco.i_evaluator import IEvaluator
+from force_bdss.api import BaseMCO, DataValue
 
 from .subprocess_workflow_evaluator import SubprocessWorkflowEvaluator
 from .scaling_tools.kpi_scaling import sen_scaling_method
+from .optimizers.optimizers import IOptimizer, WeightedOptimizer
 
 log = logging.getLogger(__name__)
-
-
-class IOptimizer(Interface):
-    def _score(self, *args, **kwargs):
-        """ Objective function score with given parameters"""
-
-    def optimize(self):
-        """ Perform an optimization procedure"""
-
-
-@provides(IOptimizer)
-class WeightedEvaluator(HasStrictTraits):
-    """Performs an optimization given a set of weights for the individual
-    KPIs.
-    """
-
-    single_point_evaluator = Instance(IEvaluator)
-
-    weights = List(Float)
-
-    parameters = List(BaseMCOParameter)
-
-    def __init__(self, single_point_evaluator, weights, parameters):
-        super(WeightedEvaluator, self).__init__(
-            single_point_evaluator=single_point_evaluator,
-            weights=weights,
-            parameters=parameters,
-        )
-
-    def _score(self, point):
-
-        score = np.dot(
-            self.weights, self.single_point_evaluator.evaluate(point)
-        )
-
-        log.info("Weighted score: {}".format(score))
-
-        return score
-
-    def optimize(self):
-        initial_point = [p.initial_value for p in self.parameters]
-        constraints = [(p.lower_bound, p.upper_bound) for p in self.parameters]
-
-        weighted_score_func = self._score
-
-        log.info("Running optimisation.")
-        log.info("Initial point: {}".format(initial_point))
-        log.info("Constraints: {}".format(constraints))
-        optimal_point = opt(weighted_score_func, initial_point, constraints)
-        optimal_kpis = self.single_point_evaluator.evaluate(optimal_point)
-        log.info("Optimal point : {}".format(optimal_point))
-        log.info("KPIs at optimal point : {}".format(optimal_kpis))
-
-        return optimal_point, optimal_kpis
 
 
 class MCO(BaseMCO):
@@ -82,7 +19,7 @@ class MCO(BaseMCO):
     optimizer = Type(IOptimizer)
 
     def _optimizer_default(self):
-        return WeightedEvaluator
+        return WeightedOptimizer
 
     def get_scaling_factors(
         self, evaluator, kpis, parameters, scaling_method=None
@@ -100,9 +37,11 @@ class MCO(BaseMCO):
             Instance that provides optimization functionality
         kpis: List[KPISpecification]
             List of KPI objects to scale
+        parameters: List[MCOParameters]
+            MCO parameters, required by the optimizer
         scaling_method: callable
             A method to scale KPI weights. Default set to the Sen's
-            "Multi-Objective Programming Method".
+            "Multi-Objective Programming Method"
         """
         if scaling_method is None:
             scaling_method = sen_scaling_method
@@ -179,12 +118,3 @@ class MCO(BaseMCO):
                 [DataValue(value=v) for v in optimal_kpis],
                 scaled_weights,
             )
-
-
-def opt(weighted_score_func, initial_point, constraints):
-    """Partial func. Performs a scipy optimise with SLSQP method given the
-    scoring function, the initial point, and a set of constraints."""
-
-    return optimize.minimize(
-        weighted_score_func, initial_point, method="SLSQP", bounds=constraints
-    ).x
